Update podscriber.py
allright...files are being deleted. that was unexciting

tldr It appears that the issue is due to an exception occurring in the add_podcast_to_db_chroma function, which prevents the new_files.append(...) line from being executed. As a result, the new_files list remains empty, and the MP3 and WAV files are not deleted at the end of your script.

In your add_podcast_to_db_chroma function, there's a print statement that references transcript_github_url, which is not defined within the scope of that function. This leads to a NameError exception.